Varpred means

Compute the mean of central estimates or the focal predictor

Usage

# S3 method for varpred
getmeans(
  object,
  what = c("both", "estimate", "focal"),
  focal = NULL,
  modelname = NULL
)

Arguments

object

varpred object.

what

a character specifying what mean to compute. If "estimate", the mean of the central estimate is computed, if "focal" the mean of the focal predictor, otherwise, both are computed.

focal

a character specifying the name of focal predictor. If NULL (default), x.var from the object is used.

modelname

character string naming varpred objects. Useful when comparing several objects.

Value

a data frame.

Details

Provides a quick way to compare observed data marginal means with that of the central estimate. Current version ignores interactions and averages the fit column only.

Examples

library(varpred)
library(ggplot2)
## Set theme for plots
varpredtheme()
## Fit the model
mod <- lm(mpg ~ wt + hp, mtcars)
## Effect
ef <- varpred(mod, "wt")
head(ef, 3)
#>         wt      fit        se      lwr      upr
#> 1 1.513000 26.69942 1.0742217 24.50239 28.89645
#> 2 1.638571 26.21247 0.9947684 24.17794 28.24700
#> 3 1.856429 25.36766 0.8569229 23.61506 27.12026
tail(ef, 3)
#>          wt      fit           se       lwr      upr
#> 27 5.334821 11.87904 1.343973e+00  9.130309 14.62777
#> 28 5.424000 11.53322 1.400399e+00  8.669084 14.39736
#> 29 3.210747 20.11584 1.404951e-15 20.115841 20.11584

## Compute means of the predictions
ef_m <- getmeans(ef, what="both", modelname="estimated")
print(ef_m)
#>         wt      fit     model
#> 1 3.210747 20.11584 estimated
# Data mean
mpg_m <- mean(mtcars$mpg)
print(mpg_m)
#> [1] 20.09062